지도 이동가능한 마커 + 주소검색

+++ PleaseSunshine/Models/CostData.swift
@@ -15,6 +15,6 @@ struct CostData: Codable {
 
 struct Cost: Codable {
     let watt, savedMoney, installCostAvg: Int
-    let bePoint: Double
+    let bePoint: Int
     let volunteer, coffee: Int
 }
